requirements.txt
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langgraph==0.2.28
langchain-groq==0.1.9
chromadb==0.5.5
sentence-transformers==3.0.1
openai==1.43.0
python-dotenv==1.0.1
pydantic==2.8.2
